pyproject.toml
[project]
name = "trello-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "fastmcp==2.5.1",
]
